Adds implementation of keycloak-js
